Fix Mac compile problems in rexxutil
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1998 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ rexutils/unix/rexxutil.cpp
@@ -3469,7 +3469,7 @@ APIRET APIENTRY SysRequestMutexSem(
 
     if(!getval(apidata->rexxutilsems, handle)){/* if sem is locked     */
                                                /* and I'm the owner    */
-      if((opencnt[handle][1]) == (int)pthread_self()){
+      if((opencnt[handle][1]) == pthread_self()){
         sprintf(retstr->strptr, "%d", 0);    /* no errors              */
       }
       else {                             /* wait for it                */
@@ -3491,7 +3491,7 @@ APIRET APIENTRY SysRequestMutexSem(
   else {                                 /* need timeout porcess       */
       if(!getval(apidata->rexxutilsems, handle)){/* if sem is locked   */
                                                  /* and I'm the owner  */
-        if((opencnt[handle][1])== (int)pthread_self()){
+        if((opencnt[handle][1]) == pthread_self()){
           sprintf(retstr->strptr, "%d", 0);/* no errors                */
         }
         else {                           /* wait for it                */
@@ -3606,7 +3606,7 @@ APIRET APIENTRY SysReleaseMutexSem(
   }
   if(!getval(apidata->rexxutilsems, handle)){/* if sem is locked     */
                                              /* and I'm the owner    */
-    if((opencnt[handle][1])== (int)pthread_self()){
+    if((opencnt[handle][1])== pthread_self()){
       /* unlock the sem                                              */
       unlocksem(apidata->rexxutilsems, handle);
       (opencnt[handle][1])= 0;         /* reset the owner TID        */
